refactor(otel): move tracing middleware to otel package and remove old tracing code
- Remove all code under hyperf/jet/middleware/tracing, including tracing.go, config.go,
  semconv builder and attributes, tests, and documentation
- Add semconv attribute utilities to hyperf/jet/middleware/otel/internal/semconv for reuse
- Update codecov and versions.yaml to replace tracing middleware references with otel
- Maintain consistency in semantic conventions by using OpenTelemetry semconv package
- Prepare for new OpenTelemetry based middleware under otel package replacing legacy tracing module
