+++ codecov.yml
@@ -66,7 +66,7 @@ fixes:
   - github.com/go-fries/fries/hyperf/jet/middleware/recovery/v3::hyperf/jet/middleware/recovery/
   - github.com/go-fries/fries/hyperf/jet/middleware/retry/v3::hyperf/jet/middleware/retry/
   - github.com/go-fries/fries/hyperf/jet/middleware/timeout/v3::hyperf/jet/middleware/timeout/
-  - github.com/go-fries/fries/hyperf/jet/middleware/tracing/v3::hyperf/jet/middleware/tracing/
+  - github.com/go-fries/fries/hyperf/jet/middleware/otel/v3::hyperf/jet/middleware/otel/
   - github.com/go-fries/fries/kratos/log/otel/v3::kratos/log/otel/
   - github.com/go-fries/fries/kratos/log/stack/v3::kratos/log/stack/
   - github.com/go-fries/fries/kratos/log/syslog/v3::kratos/log/syslog/
