+++ hyperf/jet/middleware/otel/config.go
@@ -1,4 +1,4 @@
-package tracing
+package otel
 
 import (
 	"go.opentelemetry.io/otel"
