+++ hyperf/jet/middleware/otel/version_test.go
@@ -1,4 +1,4 @@
-package tracing_test
+package otel_test
 
 import (
 	"regexp"
@@ -6,7 +6,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/go-fries/fries/hyperf/jet/middleware/tracing/v3"
+	"github.com/go-fries/fries/hyperf/jet/middleware/otel/v3"
 )
 
 // regex taken from https://semver.org/#is-there-a-suggested-regular-expression-regex-to-check-a-semver-string
@@ -16,6 +16,6 @@ var versionRegex = regexp.MustCompile(`^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)
 	`(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$`)
 
 func TestVersionSemver(t *testing.T) {
-	v := tracing.Version()
+	v := otel.Version()
 	assert.NotNil(t, versionRegex.FindStringSubmatch(v), "version is not semver: %s", v)
 }
