+++ versions.yaml
@@ -76,7 +76,7 @@ module-sets:
       - github.com/go-fries/fries/hyperf/jet/middleware/recovery/v3
       - github.com/go-fries/fries/hyperf/jet/middleware/retry/v3
       - github.com/go-fries/fries/hyperf/jet/middleware/timeout/v3
-      - github.com/go-fries/fries/hyperf/jet/middleware/tracing/v3
+      - github.com/go-fries/fries/hyperf/jet/middleware/otel/v3
 
       # Hashing
       - github.com/go-fries/fries/hashing/v3
